file/search moved to tools/media
